+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tests for heal package.
+"""